+++ post-expirator.php
@@ -5,7 +5,7 @@
  * Plugin URI: http://wordpress.org/extend/plugins/post-expirator/
  * Description: PublishPress Future allows you to schedule automatic changes to posts, pages and other content types.
  * Author: PublishPress
- * Version: 4.1.1
+ * Version: 4.1.2
  * Author URI: http://publishpress.com
  * Text Domain: post-expirator
  * Domain Path: /languages
@@ -27,7 +27,7 @@ defined('ABSPATH') or die('Direct access not allowed.');
 global $wp_version;
 
 $min_php_version = '7.4';
-$min_wp_version  = '6.7-RC4';
+$min_wp_version  = '6.7';
 
 // If the PHP or WP version is not compatible, terminate the plugin execution.
 $invalid_php_version = version_compare(phpversion(), $min_php_version, '<');
@@ -51,7 +51,7 @@ if (! defined('PUBLISHPRESS_FUTURE_LOADED')) {
         }
 
         if (! defined('PUBLISHPRESS_FUTURE_VERSION')) {
-            define('PUBLISHPRESS_FUTURE_VERSION', '4.1.1');
+            define('PUBLISHPRESS_FUTURE_VERSION', '4.1.2');
         }
 
         if (! defined('PUBLISHPRESS_FUTURE_PLUGIN_FILE')) {
@@ -122,6 +122,10 @@ if (! defined('PUBLISHPRESS_FUTURE_LOADED')) {
 
         HooksFacade::registerActivationHook(__FILE__, __NAMESPACE__ . '\\install');
         HooksFacade::registerDeactivationHook(__FILE__, __NAMESPACE__ . '\\uninstall');
+
+        add_action('plugins_loaded', function () {
+            load_plugin_textdomain('post-expirator', false, basename(dirname(__FILE__)) . '/languages/');
+        });
 
         add_action('init', function () {
             try {