+++ src/Modules/Settings/Controllers/Controller.php
@@ -106,6 +106,12 @@ class Controller implements InitializableInterface
             [$this, 'initMigrations'],
             20
         );
+        $this->hooks->addFilter(
+            CoreAbstractHooks::FILTER_ADMIN_TITLE,
+            [$this, 'onFilterAdminTitle'],
+            10,
+            2
+        );
     }
 
     public function onActionActivatePlugin()
@@ -216,8 +222,7 @@ class Controller implements InitializableInterface
 
                 // translators: %1$s is the link to the PHP strtotime function documentation, %2$s and %3$s are the opening and closing code tags. Please, do not translate the date format text, since PHP will not be able to calculate using non-english terms.
                 $fieldDefaultDateTimeOffsetDescription = esc_html__(
-                    'Set the offset to use for the default action date and time. For information on formatting, see %1$s
-                    . For example, you could enter %2$s+1 month%3$s or %2$s+1 week 2 days 4 hours 2 seconds%3$s or %2$snext Thursday%3$s. Please, use only terms in English.',
+                    'Set the offset to use for the default action date and time. For information on formatting, see %1$s. For example, you could enter %2$s+1 month%3$s or %2$s+1 week 2 days 4 hours 2 seconds%3$s or %2$snext Thursday%3$s. Please, use only terms in English.',
                     'post-expirator'
                 );
 
@@ -587,5 +592,14 @@ class Controller implements InitializableInterface
         }
 
         $this->settings->setScheduledWorkflowStepsCleanupRetention($stepScheduleCleanupRetention);
+    }
+
+    public function onFilterAdminTitle($adminTitle, $title)
+    {
+        if (isset($_GET['page']) && $_GET['page'] === 'publishpress-future') {
+            return str_replace($title, __('Action Settings', 'post-expirator'), $adminTitle);
+        }
+
+        return $adminTitle;
     }
 }