+++ src/Modules/Workflows/Controllers/WorkflowsList.php
@@ -403,6 +403,10 @@ class WorkflowsList implements InitializableInterface
 
     public function fixWorkflowEditorPageTitle()
     {
+        if (!isset($_GET['page']) || 'future_workflow_editor' !== $_GET['page']) {
+            return;
+        }
+
         global $title;
 
         $title = __("Action Workflow Editor", "post-expirator");